fdo#36876 WW8 filter: fix export of table margins
Actually no testcase, as the export bug does not affect our import
filter. The problem was that we wrote SPRM's, which are properties of
floating tables, so once the user edited the resulting document in Word
and imported it back to Writer, frames around tables got created.  Fix
this by only writing floating table SPRM's for tables which are already
in a fly frame.

Regression from cbba191cf0fff148935f24d111cb124bd5a5328b.

Change-Id: Ifeae93100ca80dcde34fc79f69a91064d44ea8f4