Don't export a variable called PATH_SEPARATOR
It used to be mis-spelled PATH_SEPERATOR. Now, after correcting the spelling,
it breaks the build of 3rd-party libraries using autoconfigury on Windows. The
value of PATH_SEPARATOR is ";" on Windows, and the configure scripts of those
libs run in Cygwin and assume that a PATH_SEPARATOR environment variable
should be correct for Cygwin. Which ";" isn't.

Rename the variable to LIBO_PATH_SEPARATOR.

Change-Id: Iff0eb93bf11d9f844a28be7ea4456e4c0ae10844