Let's assume nobody needs the sbase, scalc etc scripts on OS X
It's always possible to run the soffice executable from the same
location with a --base, --calc etc command-line option instead.

(That is not the normal way for Mac users to open an empty document
for the specific aspect of LibreOffice (Calc, Writer, etc) either.
Instead, the usual way (says cloph) is to keep a corresponding
template document in the Dock and click that.)

Change-Id: Iceea270c03a63f674e7dfdfa95117cb3bdd15184
Reviewed-on: https://gerrit.libreoffice.org/3930
Reviewed-by: Fridrich Strba <fridrich@documentfoundation.org>
Tested-by: Fridrich Strba <fridrich@documentfoundation.org>