Fix Core Text GetCaretPositions()
The secondary caret is a special caret that is inserted when the text
changes its direction e.g. between an RTL and LTR segments, not
whatever who wrote this code thought it is.

This should now be more or less the same as ATSUI version (for better or
worse), though it probably makes no difference anyway since
GetCaretPositions(), despite its name, is *not* used for determining
caret positions but only for drawing mnemonic underlines, and we don’t
draw any menus by ourselves on Mac.

While at it, adopt variable naming used in the rest of the code (not the
spacing, though. Why any sane person would want no space before opening
parenthesis and space after it!).

Change-Id: I3e8d1db33c899d0c69f65b57f0a52d10cbed1025