SfxLibraryContainer::storeLibraries_Impl: make saving basic macros work again
Regression from 6e7cdf1ac0613fa74f889ea43d9dd87ee7c9bc59, that commit
accidentally not only removed the #if 0 block, but also two other
braces -- because of this, in case bStorage was false, nothing was
saved. The effect of this was that if the user edited some macro in the
basic IDE, then saved the macro, it was shown as saved, but in fact that
didn't happen.

Change-Id: I20f8358dddfb226976be72f95dcad64de88d83c3