bnc#802888 SwPageFrm: one more clickable field fix
This is similar to 03343a817a45cc194fb52875e627efd7e4b95645. If we have
both a background object and some clickable field to select, we should
always prefer the clickable field.

This time, the problem was to click on a clickable field, when the
cursor was in normal text and we had a full-page background object.

Change-Id: Id25036d8cb2ddf62011c22bfc50844b793b942bd