Revert "fix gold warning about hidden symbols"
This reverts commit aa8e03826f9823e01fd7f7be70ec1f7c84b01444.  Whatever the gold
warnings are exactly, this is apparently not the right fix.

The change to openssl/ExternalProject_openssl.mk caused crashes when LO's
libneon.so, statically linking in LO's libssl.a, started to export symbols like
SSLv23_client_method as global, competing with the same symbol from a system's
libssl.so.  See the mail thread starting at
<http://lists.freedesktop.org/archives/libreoffice/2013-June/053119.html>
"strange crash in SSL code while opening some documents."

Whether the change to expat/StaticLibrary_expat.mk caused similar problems is
not clear to me, but it is probably best to revert both to be on the save side,
and see what exactly the gold warning was about and how to fix it the right way
if it ever shows up again.