Disable Graphite for Android and iOS
It definitely does not make sense for iOS, I assume. For Android the situation
is unclear, but let's disable it for now. It causes linking errors currently
anyway as the graphite_serverfont.cxx is not compiled for Android. (Whether it
could and should be compiled then instead of disabling Graphite, I don't
know.)

Change-Id: I1a874d304af508d2217da08e49dc158664f2e9d2