fix for bnc#823935
window network drive urls on import get internall converted to SMB
(so it appears anyway) when exported the file:/// that was on
the target ( in the relId ) for the hyperlink is dropped. Apparently
the scheme isn't necessary for the binary import. If one of these
strange 'detected as smb' window shares are exported make sure we
add the file:///
(cherry picked from commit 10ca3b49904394e1bc53ce26a45ec98b169638d7)

Change-Id: Id62e1f4845bfbc610f73499881dc33d3d955923b