GSOC work week 3, showing methods in a ListBox
This patch allows the Code Completition feature to list methods in a custom ListBox class called CodeCompleteListBox.
So, when the user presses the dot("."), a ListBox appears, and listed the methods(not just prints on the terminal).
The user can select one from them, and it is put in the source code (after the dot).

Change-Id: Ie5165e7bdaae1d96bbf40a9b996ca8ebbdb40dea