fdo#43765, fdo#57884, fdo#58052, fdo#63949: disappearing form controls
Transparent form controls in Writer may not be painted properly, if
painted from ViewShell::ImplUnlockPaint(true), called from
SwLayIdle::SwLayIdle.

The problem is that SdrPaintWindow::OutputPreRenderDevice will paint
the background of the transparent form controls (thus over-painting the
controls), so it has to be followed by painting the controls themselves.

Change-Id: Icda4ad835a398bbd50139be32ff5014a34f23bd5