fdo#65123 Revert "n#782833: (PPT) Fix font size at end of para."
This reverts commit d737b58a427e2dac47dca7d09361d266c179f8b5.

The description in the orig. commit: "Font size at the end of the para
when there is a new(empty) line is set using an extra charprops
definition." That would mean the extra StyleTextPropAtom record does not
have a record header, which seems pretty fishy. I suspect the real
problem is somewhere else (or the ppt is broken). But I cannot check it,
because I cannot access the bug.

Conflicts:
	filter/source/msfilter/svdfppt.cxx

Change-Id: I1d555cd2905ef8f297d1a96f56bebbea8f83ad62