allow these cond formats to be updated when formula val changes,fdo#64401
When the formulas depend on cells outside of the paint range we need to
update the whole range. This approach will update the range more often
than necessary but prevents that we need to mess with ScFormulaCell.

In the future we should replace ScFormulaCell in this part of the code
with something simpler that will correctly fix this problem.

Change-Id: Ie8a8e5205f9ea1e43bb6d9f8bee6e0bab29e4840