fdo#66565 RTF import: don't produce cell widths smaller than MINLAY
In case we have a table of a given width and the second (or later) row
has fewer cells, we have to add a fake cell to such a row. However, it
doesn't make sense to do this when the difference is only a few twips:
we can't create such a small frame inside the cell later anyway.

Regression from 02297f0fc12e47a0644adebecbb8622a5532015e.

Change-Id: Ibc0f02d4184b58bd423c3405e786e1ec25b9dd13